requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
cffi==1.10.0
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
Django==1.11.2
django-appconf==1.0.2
django-bootstrap3==8.2.3
django-compressor==2.1.1
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.5
incremental==17.5.0
lxml==3.8.0
nose==1.3.7
packaging==16.8
parsel==1.2.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==17.0.0
pyparsing==2.2.0
pytz==2017.2
queuelib==1.4.2
rcssmin==1.0.6
rjsmin==1.0.12
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
Twisted==17.1.0
w3lib==1.17.0
zope.interface==4.4.1
